beta 2.10.1 Changed some competition gui layout To terminate a competition, now use "/competition terminate" instead of "/competition end" Changed some translation key and text Fixed some item lore showing incorrect text Various fixed and tweaks
