+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 group = 'org.macausmp'
-version = '2.10-BETA'
+version = '2.10.1-BETA'
 
 repositories {
     mavenCentral()
