+++ src/main/java/org/macausmp/sportsday/competition/IceBoatRacing.java
@@ -30,6 +30,7 @@ public class IceBoatRacing extends AbstractTrackEvent {
     @Override
     protected void onSetup() {
         getLocation().getWorld().getEntitiesByClass(Boat.class).forEach(Boat::remove);
+        boatMap.clear();
         getContestants().forEach(data -> boatMap.put(data.getUUID(), boat(data.getPlayer())));
     }
 
@@ -37,7 +38,11 @@ public class IceBoatRacing extends AbstractTrackEvent {
     protected void onStart() {}
 
     @Override
-    protected void onEnd(boolean force) {
+    protected void onEnd() {}
+
+    @Override
+    protected void cleanup() {
+        super.cleanup();
         boatMap.values().forEach(Boat::remove);
     }
