+++ src/main/java/org/macausmp/sportsday/gui/competition/AbstractCompetitionGUI.java
@@ -13,9 +13,9 @@ import org.bukkit.persistence.PersistentDataType;
 import org.jetbrains.annotations.NotNull;
 import org.macausmp.sportsday.SportsDay;
 import org.macausmp.sportsday.competition.*;
-import org.macausmp.sportsday.competition.sumo.Sumo;
 import org.macausmp.sportsday.gui.ButtonHandler;
 import org.macausmp.sportsday.gui.PluginGUI;
+import org.macausmp.sportsday.gui.competition.event.EventGUI;
 import org.macausmp.sportsday.gui.competition.event.JavelinGUI;
 import org.macausmp.sportsday.gui.competition.event.SumoGUI;
 import org.macausmp.sportsday.gui.competition.event.TrackEventGUI;
@@ -66,18 +66,18 @@ public abstract class AbstractCompetitionGUI extends PluginGUI {
     public void console(@NotNull InventoryClickEvent e, @NotNull Player p, @NotNull ItemStack item) {
         p.playSound(Sound.sound(Key.key("minecraft:ui.button.click"), Sound.Source.MASTER, 1f, 1f));
         IEvent event = Competitions.getCurrentEvent();
-        if (event != null && event.getStatus() != Status.ENDED) {
-            if (event == Competitions.JAVELIN_THROW)
-                p.openInventory(new JavelinGUI((JavelinThrow) event).getInventory());
-            else if (event == Competitions.SUMO)
-                p.openInventory(new SumoGUI((Sumo) event).getInventory());
-            else if (event instanceof ITrackEvent track)
-                p.openInventory(new TrackEventGUI(track).getInventory());
-            else
-                p.openInventory(new CompetitionConsoleGUI().getInventory());
-        } else {
+        if (event == null || event.getStatus() == Status.ENDED) {
             p.openInventory(new CompetitionConsoleGUI().getInventory());
+            return;
         }
+        if (event == Competitions.JAVELIN_THROW)
+            p.openInventory(new JavelinGUI((JavelinThrow) event).getInventory());
+        else if (event == Competitions.SUMO)
+            p.openInventory(new SumoGUI((Sumo) event).getInventory());
+        else if (event instanceof ITrackEvent track)
+            p.openInventory(new TrackEventGUI(track).getInventory());
+        else
+            p.openInventory(new EventGUI<>(event).getInventory());
     }
 
     @ButtonHandler("contestants_list")
