+++ src/main/java/org/macausmp/sportsday/gui/competition/CompetitionConsoleGUI.java
@@ -4,24 +4,18 @@ import net.kyori.adventure.key.Key;
 import net.kyori.adventure.sound.Sound;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
-import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.macausmp.sportsday.competition.Competitions;
-import org.macausmp.sportsday.competition.IEvent;
 import org.macausmp.sportsday.competition.Status;
 import org.macausmp.sportsday.gui.ButtonHandler;
-import org.macausmp.sportsday.gui.ConfirmationGUI;
 import org.macausmp.sportsday.util.ItemUtil;
 
 public class CompetitionConsoleGUI extends AbstractCompetitionGUI {
     private static final ItemStack START_COMPETITION = ItemUtil.head(ItemUtil.START, "start_competition", "gui.start.title", "gui.start.lore");
     private static final ItemStack LOAD_COMPETITION = ItemUtil.head(ItemUtil.START, "load_competition", "gui.load.title", "gui.load.lore");
-    private static final ItemStack PAUSE_COMPETITION = ItemUtil.head(ItemUtil.PAUSE, "pause_competition", "gui.pause.title", "gui.pause.lore");
-    private static final ItemStack UNPAUSE_COMPETITION = ItemUtil.head(ItemUtil.START, "unpause_competition", "gui.unpause.title", "gui.unpause.lore");
-    private static final ItemStack END_COMPETITION = ItemUtil.item(Material.RED_CONCRETE, "end_competition", "gui.end.title", "gui.end.lore");
 
     public CompetitionConsoleGUI() {
         super(27, Component.translatable("gui.console.title"));
@@ -31,31 +25,8 @@ public class CompetitionConsoleGUI extends AbstractCompetitionGUI {
         getInventory().setItem(1, CONTESTANTS_LIST);
         getInventory().setItem(2, COMPETITION_SETTINGS);
         getInventory().setItem(3, VERSION);
-        if (hasEvent()) {
-            getInventory().setItem(18, status());
-            getInventory().setItem(19, player());
-            getInventory().setItem(21, PAUSE_COMPETITION);
-            getInventory().setItem(22, UNPAUSE_COMPETITION);
-            getInventory().setItem(26, END_COMPETITION);
-        } else {
-            getInventory().setItem(18, START_COMPETITION);
-            getInventory().setItem(19, LOAD_COMPETITION);
-        }
-        update();
-    }
-
-    @Override
-    public void update() {
-        if (hasEvent()) {
-            getInventory().setItem(18, status());
-            getInventory().setItem(19, player());
-        }
-    }
-
-    public static void updateGUI() {
-        PLUGIN.getServer().getOnlinePlayers().stream().map(p -> p.getOpenInventory().getTopInventory())
-                .filter(inv -> inv.getHolder() instanceof CompetitionConsoleGUI)
-                .map(inv -> (CompetitionConsoleGUI) inv.getHolder()).forEach(CompetitionConsoleGUI::update);
+        getInventory().setItem(18, START_COMPETITION);
+        getInventory().setItem(19, LOAD_COMPETITION);
     }
 
     @ButtonHandler("start_competition")
@@ -78,65 +49,5 @@ public class CompetitionConsoleGUI extends AbstractCompetitionGUI {
         }
         Competitions.loadEventData(p);
         p.playSound(Sound.sound(Key.key("minecraft:ui.button.click"), Sound.Source.MASTER, 1f, 1f));
-    }
-
-    @ButtonHandler("pause_competition")
-    public void pause(@NotNull InventoryClickEvent e, @NotNull Player p, @NotNull ItemStack item) {
-        if (Competitions.getCurrentEvent() == null || Competitions.getCurrentEvent().getStatus() == Status.ENDED) {
-            p.sendMessage(Component.translatable("command.competition.null_event").color(NamedTextColor.RED));
-            p.playSound(Sound.sound(Key.key("minecraft:entity.enderman.teleport"), Sound.Source.MASTER, 1f, 1f));
-            return;
-        }
-        p.playSound(Sound.sound(Key.key(Competitions.pause(p) ?
-                "minecraft:ui.button.click" : "minecraft:entity.enderman.teleport"), Sound.Source.MASTER, 1f, 1f));
-    }
-
-    @ButtonHandler("unpause_competition")
-    public void unpause(@NotNull InventoryClickEvent e, @NotNull Player p, @NotNull ItemStack item) {
-        if (Competitions.getCurrentEvent() == null || Competitions.getCurrentEvent().getStatus() == Status.ENDED) {
-            p.sendMessage(Component.translatable("command.competition.null_event").color(NamedTextColor.RED));
-            p.playSound(Sound.sound(Key.key("minecraft:entity.enderman.teleport"), Sound.Source.MASTER, 1f, 1f));
-            return;
-        }
-        p.playSound(Sound.sound(Key.key(Competitions.unpause(p) ?
-                "minecraft:ui.button.click" : "minecraft:entity.enderman.teleport"), Sound.Source.MASTER, 1f, 1f));
-    }
-
-    @ButtonHandler("end_competition")
-    public void end(@NotNull InventoryClickEvent e, @NotNull Player p, @NotNull ItemStack item) {
-        if (Competitions.getCurrentEvent() == null || Competitions.getCurrentEvent().getStatus() == Status.ENDED) {
-            p.sendMessage(Component.translatable("command.competition.null_event").color(NamedTextColor.RED));
-            p.playSound(Sound.sound(Key.key("minecraft:entity.enderman.teleport"), Sound.Source.MASTER, 1f, 1f));
-            return;
-        }
-        p.playSound(Sound.sound(Key.key("minecraft:ui.button.click"), Sound.Source.MASTER, 1f, 1f));
-        p.openInventory(new ConfirmationGUI(this, player -> {
-            boolean b = Competitions.forceEnd(player);
-            player.playSound(Sound.sound(Key.key(b ?
-                    "minecraft:ui.button.click" : "minecraft:entity.enderman.teleport"), Sound.Source.MASTER, 1f, 1f));
-            return b;
-        }).getInventory());
-    }
-
-    private @NotNull ItemStack status() {
-        boolean b = Competitions.getCurrentEvent() != null;
-        Component display = Component.translatable("competition.current").color(NamedTextColor.GREEN)
-                .arguments(b ? Competitions.getCurrentEvent().getName() : Component.translatable("gui.text.none"));
-        Component lore = Component.translatable("competition.status").color(NamedTextColor.GREEN)
-                .arguments(b ? Competitions.getCurrentEvent().getStatus().getName() : Status.IDLE.getName());
-        return ItemUtil.item(Material.BEACON, null, display, lore);
-    }
-
-    private @NotNull ItemStack player() {
-        Component display = Component.translatable("competition.contestants.total").color(NamedTextColor.GREEN)
-                .arguments(Component.text(Competitions.getContestants().size()).color(NamedTextColor.YELLOW));
-        Component lore = Component.translatable("competition.contestants.online").color(NamedTextColor.GREEN)
-                .arguments(Component.text(Competitions.getOnlineContestants().size()).color(NamedTextColor.YELLOW));
-        return ItemUtil.item(Material.PAPER, null, display, lore);
-    }
-
-    private static boolean hasEvent() {
-        IEvent event = Competitions.getCurrentEvent();
-        return event != null && event.getStatus() != Status.ENDED;
     }
 }
