+++ src/main/java/org/macausmp/sportsday/gui/competition/CompetitionSettingsGUI.java
@@ -9,11 +9,7 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.persistence.PersistentDataType;
 import org.jetbrains.annotations.NotNull;
-import org.macausmp.sportsday.competition.Competitions;
-import org.macausmp.sportsday.competition.IEvent;
-import org.macausmp.sportsday.competition.ITrackEvent;
-import org.macausmp.sportsday.competition.JavelinThrow;
-import org.macausmp.sportsday.competition.sumo.Sumo;
+import org.macausmp.sportsday.competition.*;
 import org.macausmp.sportsday.gui.ButtonHandler;
 import org.macausmp.sportsday.gui.competition.setting.EventSettingsGUI;
 import org.macausmp.sportsday.gui.competition.setting.SumoSettingsGUI;
