+++ src/main/resources/lang/Bundle_en_US.properties
@@ -17,7 +17,7 @@ command.competition.pause.success=已暫停比賽
 command.competition.pause.failed=暫停失敗，該階段不允許暫停比賽或者比賽已被暫停
 command.competition.unpause.success=已取消暫停比賽
 command.competition.unpause.failed=取消暫停失敗，該比賽不在暫停狀態中
-command.competition.end.success=已強制結束一場比賽
+command.competition.terminate.success=已終止一場比賽
 command.competition.null_event=現在沒有比賽進行中
 command.competition.invalid_status=當前階段不能使用此操作
 command.competition.not_savable=當前比賽不是可被儲存的比賽
@@ -51,7 +51,7 @@ contestant.practice.teleport.venue=§e您已被傳送至{0}§e場地
 contestant.practice.finished=§e您完成了{0}§e訓練，您現在可以繼續訓練或者選擇退出
 console.competition.coming=一場{0}即將開始
 console.competition.end={0}已結束
-console.competition.force_end={0}已強制結束
+console.competition.terminate={0}已終止
 event.name.elytra_racing=§e飛行賽
 event.name.ice_boat_racing=§b冰船賽
 event.name.javelin_throw=§e擲標槍
@@ -102,13 +102,14 @@ event.sumo.next_stage_countdown={0}秒後進入下一個階段
 event.sumo.knockback_stick.countdown={0}秒後將派發退擊棒
 event.sumo.knockback_stick.given=§e已派發退擊棒
 event.sumo.rank=第{0}名 {1}
-event.ready.broadcast={0}將於{1}秒後開始，請選手們各就位
-event.start.countdown=比賽將於{0}秒開始
-event.start.broadcast=§a比賽開始
-event.pause.broadcast=§e比賽暫停中
-event.unpause.broadcast=§e比賽繼續
-event.end.broadcast=§c比賽結束
-broadcast.play_mvp_anthem=§e正在播放{0}§e的MVP凱歌 - {1}
+event.broadcast.ready={0}將於{1}秒後開始，請選手們各就位
+event.broadcast.start_countdown=比賽將於{0}秒開始
+event.broadcast.start=§a比賽開始
+event.broadcast.pause=§e比賽暫停中
+event.broadcast.unpause=§e比賽繼續
+event.broadcast.end=§c比賽結束
+event.broadcast.terminate=§c比賽終止
+event.broadcast.play_mvp_anthem=§e正在播放{0}§e的MVP凱歌 - {1}
 role.contestants=選手
 role.referees=裁判
 role.audiences=觀眾
@@ -144,6 +145,7 @@ gui.menu.practice.lore=§e點擊打開項目傳送菜單
 gui.menu.practice=§e點擊傳送到{0}§e場地
 gui.console.title=比賽控制面板
 gui.console.lore=§7點擊進入比賽控制面板
+gui.event_console.title={0}控制面板
 gui.competition.track.player=§e選手：{0}
 gui.competition.track.result=§e成綪：{0}
 gui.competition.track.second={0}秒
@@ -178,8 +180,8 @@ gui.pause.title=暫停比賽
 gui.pause.lore=§7點擊暫停比賽
 gui.unpause.title=取消暫停比賽
 gui.unpause.lore=§7點擊取消暫停比賽
-gui.end.title=§c結束當前比賽
-gui.end.lore=§7點擊結束當前比賽
+gui.terminate.title=§c終止當前比賽
+gui.terminate.lore=§7點擊終止當前比賽
 gui.settings.title=比賽設定
 gui.settings.lore=§7點擊查看比賽設定
 gui.event_settings.title={0}設定
@@ -187,7 +189,7 @@ gui.event_settings.lore=§7點擊查看{0}§7設定
 gui.event_settings.enable=可用狀態：{0}
 gui.event_settings.amount=最少需要的選手人數：{0}人
 gui.event_settings.location=場地位置：
-gui.event_settings.location_lore=§a點擊設置為您當前所站位置
+gui.event_settings.location_lore=§7點擊設置為您當前所站位置
 gui.event_settings.laps=完成比賽所需的圈數：{0}圈
 gui.event_settings.weapon=啟用武器：{0}
 gui.event_settings.reload_require=§e§n§o該設定需要在重新加載伺服器後才能生效
